remove dead code
