+++ packages/cli/src/index.ts
@@ -63,7 +63,6 @@ import { APIError } from './util/errors-ts';
 import { SENTRY_DSN } from './util/constants';
 import getUpdateCommand from './util/get-update-command';
 import { getCommandName, getTitleName } from './util/pkg-name';
-import doLoginPrompt from './util/login/prompt';
 import type { AuthConfig, GlobalConfig } from '@vercel-internals/types';
 import type { VercelConfig } from '@vercel/client';
 import { ProxyAgent } from 'proxy-agent';
@@ -72,11 +71,11 @@ import { execExtension } from './util/extension/exec';
 import { TelemetryEventStore } from './util/telemetry';
 import { RootTelemetryClient } from './util/telemetry/root';
 import { help } from './args';
-import { updateCurrentTeamAfterLogin } from './util/login/update-current-team-after-login';
 import { checkTelemetryStatus } from './util/telemetry/check-status';
 import output from './output-manager';
 import { checkGuidanceStatus } from './util/guidance/check-status';
 import { determineAgent } from '@vercel/detect-agent';
+import login from './commands/login';
 
 const VERCEL_DIR = getGlobalPathConfig();
 const VERCEL_CONFIG_PATH = configFiles.getConfigFilePath();
@@ -402,21 +401,9 @@ const main = async () => {
     if (isTTY) {
       output.log(`No existing credentials found. Please log in:`);
       try {
-        const result = await doLoginPrompt(client);
-
+        const result = await login(client);
         // The login function failed, so it returned an exit code
-        if (typeof result === 'number') {
-          return result;
-        }
-
-        // When `result` is a string it's the user's authentication token.
-        // It needs to be saved to the configuration file.
-        client.authConfig.token = result.token;
-
-        await updateCurrentTeamAfterLogin(client, result.teamId);
-
-        configFiles.writeToAuthConfigFile(client.authConfig);
-        configFiles.writeToConfigFile(client.config);
+        if (result !== 0) return result;
       } catch (error) {
         printError(error);
         return 1;
