+++ packages/cli/src/util/login/prompt.ts
@@ -1,70 +1,28 @@
 import type Client from '../client';
 import error from '../output/error';
-import listInput from '../input/list';
 import { getCommandName } from '../pkg-name';
-import type { LoginResult, SAMLError } from './types';
+import type { SAMLError } from './types';
 import doSamlLogin from './saml';
-import doEmailLogin from './email';
-import doGithubLogin from './github';
-import doGoogleLogin from './google';
-import doGitlabLogin from './gitlab';
-import doBitbucketLogin from './bitbucket';
 import output from '../../output-manager';
+import login from '../../commands/login';
 
 export default async function prompt(
   client: Client,
   error?: Pick<SAMLError, 'teamId' | 'scope'>,
-  outOfBand?: boolean,
   ssoUserId?: string
 ) {
-  let result: LoginResult = 1;
-
-  const choices = [
-    { name: 'Continue with GitHub', value: 'github', short: 'github' },
-    { name: 'Continue with Google', value: 'google', short: 'google' },
-    { name: 'Continue with GitLab', value: 'gitlab', short: 'gitlab' },
-    { name: 'Continue with Bitbucket', value: 'bitbucket', short: 'bitbucket' },
-    { name: 'Continue with Email', value: 'email', short: 'email' },
-    { name: 'Continue with SAML Single Sign-On', value: 'saml', short: 'saml' },
-  ];
-
-  if (ssoUserId || (error && !error.teamId)) {
-    // Remove SAML login option if we're connecting SAML Profile,
-    // or if this is a SAML error for a user / team without SAML
-    choices.pop();
-  }
-
-  const choice = await listInput(client, {
-    message: 'Log in to Vercel',
-    choices,
-  });
-
-  if (choice === 'google') {
-    result = await doGoogleLogin(client, outOfBand, ssoUserId);
-  } else if (choice === 'github') {
-    result = await doGithubLogin(client, outOfBand, ssoUserId);
-  } else if (choice === 'gitlab') {
-    result = await doGitlabLogin(client, outOfBand, ssoUserId);
-  } else if (choice === 'bitbucket') {
-    result = await doBitbucketLogin(client, outOfBand, ssoUserId);
-  } else if (choice === 'email') {
-    const email = await readInput(client, 'Enter your email address:');
-    result = await doEmailLogin(client, email, ssoUserId);
-  } else if (choice === 'saml') {
+  if (error) {
     const slug =
       error?.scope ||
       error?.teamId ||
       (await readInput(client, 'Enter your Team slug:'));
-    result = await doSamlLogin(client, slug, outOfBand, ssoUserId);
+    return await doSamlLogin(client, slug, ssoUserId);
   }
 
-  return result;
+  return await login(client);
 }
 
-export async function readInput(
-  client: Client,
-  message: string
-): Promise<string> {
+async function readInput(client: Client, message: string): Promise<string> {
   let input;
 
   while (!input) {
